qbittorrent: fix qmake path error
qbittorent was failing to build with the following error:

    checking for Qt5 qmake >= 5.2.0... not found
    configure: error: Could not find qmake

Adding qt5.qmakeHook to build inputs helps, but later generates a
zillion of link errors to boost. The reason is that qbittorrent requires
configuration through autoconf, but qmakeHook overrides configurePhase
to invoke qmake directly. That can be disabled with
dontUseQmakeConfigure, but now configure doesn't find qmake again.
Turned out, configure script doesn't search qmake in the PATH, but
rather uses QT_QMAKE environment variable; this commit sets it to point
to the qmake instance created by qmakeHook.